+++ test/transform/resource/after-delombok/ValueStaticField.java
@@ -1,0 +1,29 @@
+final class ValueStaticField {
+	private static int x;
+	private static final String PASSWORD = "Ken sent me";
+	@java.lang.SuppressWarnings("all")
+	@javax.annotation.Generated("lombok")
+	public ValueStaticField() {
+	}
+	@java.lang.Override
+	@java.lang.SuppressWarnings("all")
+	@javax.annotation.Generated("lombok")
+	public boolean equals(final java.lang.Object o) {
+		if (o == this) return true;
+		if (!(o instanceof ValueStaticField)) return false;
+		return true;
+	}
+	@java.lang.Override
+	@java.lang.SuppressWarnings("all")
+	@javax.annotation.Generated("lombok")
+	public int hashCode() {
+		int result = 1;
+		return result;
+	}
+	@java.lang.Override
+	@java.lang.SuppressWarnings("all")
+	@javax.annotation.Generated("lombok")
+	public java.lang.String toString() {
+		return "ValueStaticField()";
+	}
+}