+++ test/transform/resource/before/ValueStaticField.java
@@ -1,0 +1,5 @@
+import lombok.Value;
+@Value class ValueStaticField {
+	static int x;
+	static String PASSWORD = "Ken sent me";
+}